% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sampling.R
\name{bell_gen}
\alias{bell_gen}
\title{Generate Bell-Shaped Numeric Data}
\usage{
bell_gen(start, end, mean, sd, size, unique = TRUE)
}
\arguments{
\item{start}{The start of the range for generating random numbers.}

\item{end}{The end of the range for generating random numbers.}

\item{mean}{The mean of the normal distribution.}

\item{sd}{The standard deviation of the normal distribution.}

\item{size}{The number of samples to generate.}

\item{unique}{Logical; if TRUE, ensures the samples are unique.}
}
\value{
A vector of bell-shaped numeric data.
}
\description{
Generates numeric data following a bell curve (normal distribution).
}
\examples{
bell_gen(1, 100, 50, 10, 10)
}
