% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_gen.R
\name{birth_year}
\alias{birth_year}
\title{Calculate Birth Year}
\usage{
birth_year(x, y)
}
\arguments{
\item{x}{A numeric value or vector representing age(s).}

\item{y}{A Date object or vector representing the reference date(s).}
}
\value{
A numeric value or vector representing the birth year(s).
}
\description{
Calculates the birth year given an age and a reference year.
}
\examples{
birth_year(25, as.Date("2023-01-01"))  # Returns 1998
birth_year(c(25, 30), as.Date("2023-01-01"))  # Returns c(1998, 1993)
}
