% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversion.R
\name{cel_to_fhr}
\alias{cel_to_fhr}
\title{Convert Celsius to Fahrenheit}
\usage{
cel_to_fhr(x, digits)
}
\arguments{
\item{x}{A numeric vector of temperatures in Celsius.}

\item{digits}{The number of decimal places to include in the output (default is 1).}
}
\value{
A character vector of temperatures in Fahrenheit, formatted to the specified number of decimal places.
}
\description{
Converts a numeric vector representing temperatures in Celsius to Fahrenheit.
}
\examples{
cel_to_fhr(0, digits=1)  # Converts 0°C to "32.0"
cel_to_fhr(100, digits = 2)  # Converts 100°C to "212.00"
}
