% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_gen.R
\name{date_gen}
\alias{date_gen}
\title{Generate Random Dates}
\usage{
date_gen(start, end, size)
}
\arguments{
\item{start}{The start date as a character string in "YYYY-MM-DD" format.}

\item{end}{The end date as a character string in "YYYY-MM-DD" format.}

\item{size}{The number of random dates to generate.}
}
\value{
A vector of random Date objects.
}
\description{
Generates a vector of random dates within a specified range.
}
\examples{
date_gen("2022-01-01", "2022-12-31", 10)  # Generates 10 random dates in 2022
}
