% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_gen.R
\name{date_origin}
\alias{date_origin}
\title{Convert Numeric Date to Date Object}
\usage{
date_origin(x)
}
\arguments{
\item{x}{A numeric value or vector representing the number of days since January 1, 1970.}
}
\value{
A Date object or vector of Date objects.
}
\description{
Converts a numeric value representing the number of days since January 1, 1970, to a Date object.
}
\examples{
date_origin(0)  # Converts to "1970-01-01"
date_origin(18500)  # Converts to "2020-09-18"
}
