% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversion.R
\name{env_to_df}
\alias{env_to_df}
\title{Convert Environment Variables to Data Frame}
\usage{
env_to_df(envs)
}
\arguments{
\item{envs}{A named list of environment variables.}
}
\value{
A data frame containing the environment variables.
}
\description{
Converts environment variables into a data frame.
}
\examples{
env_vars <- list(my.env.var1 = 10, my.env.var2 = 20)
env_to_df(env_vars)  # Converts the environment variables to a data frame
}
