% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sampling.R
\name{float_bell_gen}
\alias{float_bell_gen}
\title{Generate Bell-Shaped Floating Point Data}
\usage{
float_bell_gen(start, end, mean = 40, sd = 10, size, unique = TRUE, digits = 1)
}
\arguments{
\item{start}{The start of the range for generating random numbers.}

\item{end}{The end of the range for generating random numbers.}

\item{mean}{The mean of the normal distribution.}

\item{sd}{The standard deviation of the normal distribution.}

\item{size}{The number of samples to generate.}

\item{unique}{Logical; if TRUE, ensures the samples are unique.}

\item{digits}{The number of decimal places to retain.}
}
\value{
A vector of bell-shaped floating-point data.
}
\description{
Generates floating-point data following a bell curve (normal distribution).
}
\examples{
float_bell_gen(1, 100, 50, 10, 10, 2)
}
