% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_gen.R
\name{gen_after_dt}
\alias{gen_after_dt}
\title{Generate Date-Time After Given Date}
\usage{
gen_after_dt(date, range, size, replace)
}
\arguments{
\item{date}{A POSIXct date-time object.}

\item{range}{A numeric vector representing minute ranges for generating times.}

\item{size}{The number of date-times to generate.}

\item{replace}{A logical value indicating whether to allow replacement.}
}
\value{
A vector of POSIXct date-time objects.
}
\description{
Generates random date-times after a given date and time.
}
\examples{
gen_after_dt(Sys.time(), c(10, 20, 30), 5, TRUE)
# Generates 5 date-times 10, 20, or 30 minutes after the current time
}
