% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_gen.R
\name{gen_date_range}
\alias{gen_date_range}
\title{Generate Random Date-Time Range}
\usage{
gen_date_range(start, end, by)
}
\arguments{
\item{start}{The start date-time as a character string in "YYYY-MM-DD HH:MM:SS" format.}

\item{end}{The end date-time as a character string in "YYYY-MM-DD HH:MM:SS" format.}

\item{by}{The increment of the sequence (e.g., "1 day", "1 hour").}
}
\value{
A vector of random POSIXct date-time objects.
}
\description{
Generates a vector of random date-times within a specified range.
}
\examples{
gen_date_range("2023-01-01 00:00:00", "2023-01-02 00:00:00", "1 hour")
# Generates random date-times every hour from Jan 1 to Jan 2, 2023
}
