% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_gen.R
\name{gen_date_time}
\alias{gen_date_time}
\title{Generate Random Date-Time Within Hours}
\usage{
gen_date_time(date, stime, etime, size, min, replace)
}
\arguments{
\item{date}{A vector of POSIXct date-time objects.}

\item{stime}{The start time as an integer (hour in 24-hour format).}

\item{etime}{The end time as an integer (hour in 24-hour format).}

\item{size}{The number of date-times to generate.}

\item{min}{A vector of minute increments for the date-times.}

\item{replace}{A logical value indicating whether to allow replacement.}
}
\value{
A vector of POSIXct date-time objects.
}
\description{
Generates random date-times within a specific time range on a given date.
}
\examples{
gen_date_time(Sys.Date(), 9, 17, 10, c(0, 15, 30, 45), TRUE)
# Generates 10 random date-times between 9:00 and 17:00
}
