% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_gen.R
\name{gen_day_after_dt}
\alias{gen_day_after_dt}
\title{Generate Date-Time Days After Given Date}
\usage{
gen_day_after_dt(date, days, range, size, replace)
}
\arguments{
\item{date}{A POSIXct date-time object.}

\item{days}{A numeric vector representing the number of days after the base date.}

\item{range}{A numeric vector representing minute ranges for generating times.}

\item{size}{The number of date-times to generate.}

\item{replace}{A logical value indicating whether to allow replacement.}
}
\value{
A vector of POSIXct date-time objects.
}
\description{
Generates random date-times a specific number of days after a given date.
}
\examples{
gen_day_after_dt(Sys.time(), c(1, 2, 3), c(10, 20, 30), 5, TRUE)
# Generates 5 date-times 1, 2, or 3 days and 10, 20, or 30 minutes after the current time
}
