% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_gen.R
\name{gen_dt}
\alias{gen_dt}
\title{Generate Random Date-Time}
\usage{
gen_dt(date = Sys.Date())
}
\arguments{
\item{date}{The base date as a Date object. Defaults to the current date.}
}
\value{
A vector of POSIXct date-time objects.
}
\description{
Generates random date-times around the current date.
}
\examples{
gen_dt()  # Generates 1000 random date-times around the current date
}
