% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_transform.R
\name{gen_folder}
\alias{gen_folder}
\title{Generate Folder Data}
\usage{
gen_folder(crfpagename, foldername, folder, size = size)
}
\arguments{
\item{crfpagename}{The name of the CRF page.}

\item{foldername}{The name of the folder.}

\item{folder}{The folder identifier.}
}
\value{
A data frame with folder and identifier information.
}
\description{
Generates a data frame with folder information and identifiers.
}
\examples{
gen_folder("CRFPage1", "Folder1", "F1")
}
