% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_transform.R
\name{gen_visit_days}
\alias{gen_visit_days}
\title{Generate Visit Days}
\usage{
gen_visit_days(var, sdate, scn_days, size)
}
\arguments{
\item{var}{A character vector indicating visit types.}

\item{sdate}{The start date for visits.}

\item{scn_days}{The number of days from the start date for screening visits.}

\item{size}{The number of visits to generate.}
}
\value{
A vector of visit dates.
}
\description{
Generates visit days based on screening dates and specified day ranges.
}
\examples{
gen_visit_days(c("SCN", "FUP"), Sys.Date(), 7, 10)
}
