% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversion.R
\name{get_alpha}
\alias{get_alpha}
\title{Remove Digits from String}
\usage{
get_alpha(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
A character vector with all numeric characters removed.
}
\description{
Removes all numeric characters from a string.
}
\examples{
get_alpha("abc123")  # Returns "abc"
get_alpha("R2D2")  # Returns "RD"
}
