% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversion.R
\name{get_char}
\alias{get_char}
\title{Convert Data to Character}
\usage{
get_char(data)
}
\arguments{
\item{data}{The input data to be converted to a character vector.}
}
\value{
A character vector.
}
\description{
Converts any given data to a character vector.
}
\examples{
get_char(1)  # Converts the number 1 to the character "1"
get_char(c(1, 2, 3))  # Converts a numeric vector to a character vector
}
