% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_transform.R
\name{get_fcode}
\alias{get_fcode}
\title{Generate First Character Code}
\usage{
get_fcode(x)
}
\arguments{
\item{x}{A character vector to process.}
}
\value{
A character vector with the generated codes.
}
\description{
Generates a code from the first character and numeric part of a string.
}
\examples{
get_fcode(c("A123", "B456"))
}
