% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_gen.R
\name{hh_mm}
\alias{hh_mm}
\title{Extract Time from Date}
\usage{
hh_mm(x)
}
\arguments{
\item{x}{A Date object or vector of Date objects.}
}
\value{
A character value or vector representing the time(s).
}
\description{
Extracts the hour and minute from a Date object.
}
\examples{
hh_mm(as.POSIXct("2023-01-01 14:45:00"))  # Returns 14:45
}
