% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_transform.R
\name{high_low}
\alias{high_low}
\title{Categorize Values as High, Low, or Normal}
\usage{
high_low(var, high, low)
}
\arguments{
\item{var}{A numeric vector to categorize.}

\item{high}{The threshold for high values.}

\item{low}{The threshold for low values.}
}
\value{
A character vector with categories ("+", "-", "0").
}
\description{
Categorizes values as high, low, or normal based on provided thresholds.
}
\examples{
high_low(c(1, 5, 10), 8, 3)
}
