% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{lb_high}
\alias{lb_high}
\title{Compute Upper Bound}
\usage{
lb_high(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The computed upper bound.
}
\description{
Computes the upper bound as mean plus one standard deviation.
}
\examples{
lb_high(c(1, 2, 3, 4, 5))
}
