% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sampling.R
\name{list_gen}
\alias{list_gen}
\title{Generate List Elements}
\usage{
list_gen(list, size, replace = TRUE)
}
\arguments{
\item{list}{A list of elements to sample from.}

\item{size}{The number of elements to sample.}

\item{replace}{Logical; if TRUE, sampling is with replacement.}
}
\value{
A sampled list of elements.
}
\description{
Generates a sample list of elements from a given list with or without replacement.
}
\examples{
list_gen(c("apple", "banana", "cherry"), 5)
}
