% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_transform.R
\name{no_brac_txt}
\alias{no_brac_txt}
\title{Remove Text Inside Parentheses}
\usage{
no_brac_txt(x)
}
\arguments{
\item{x}{A character vector to process.}
}
\value{
A character vector with text inside parentheses removed.
}
\description{
Removes any text within parentheses in a string.
}
\examples{
no_brac_txt("Example (text) string")
}
