% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sampling.R
\name{prob_gen}
\alias{prob_gen}
\title{Generate List Elements with Probabilities}
\usage{
prob_gen(list, size, prob, replace = TRUE)
}
\arguments{
\item{list}{A list of elements to sample from.}

\item{size}{The number of elements to sample.}

\item{prob}{A vector of probabilities for each element.}

\item{replace}{Logical; if TRUE, sampling is with replacement.}
}
\value{
A sampled list of elements.
}
\description{
Generates a sample list of elements from a given list with specified probabilities.
}
\examples{
prob_gen(c("A", "B", "C"), 10, c(0.2, 0.5, 0.3))
}
