% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sampling.R
\name{rep_each}
\alias{rep_each}
\title{Repeat Elements Randomly}
\usage{
rep_each(max_size, subj, size = size)
}
\arguments{
\item{max_size}{The maximum number of repetitions for each element.}

\item{subj}{A vector of subjects from which elements are selected.}

\item{size}{The number of times to sample.}
}
\value{
A sorted vector with repeated elements.
}
\description{
Generates a sorted vector by repeating elements from a given set a random number of times up to a specified maximum.
}
\examples{
rep_each(5, c("A", "B", "C"), 10)
}
