% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sampling.R
\name{std_gen}
\alias{std_gen}
\title{Standardize Data Generation}
\usage{
std_gen(list)
}
\arguments{
\item{list}{A list containing 'value' and 'factor' elements for generating data.}
}
\value{
A data frame with standardized codes and decodes.
}
\description{
Generates standardized codes and decodes based on the given list.
}
\examples{
std_gen(list(value = c("A", "B", "C"), factor = c(0.3, 0.4, 0.3)))
}
