% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_transform.R
\name{ymd_gen}
\alias{ymd_gen}
\title{Generate Year, Month, and Day from Date}
\usage{
ymd_gen(dat)
}
\arguments{
\item{dat}{A vector of dates.}
}
\value{
A data frame with year, month, and day columns.
}
\description{
Extracts the year, month, and day from a given date and returns a data frame.
}
\examples{
ymd_gen(Sys.Date())
}
